ml-service/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-service"
version = "0.1.0"
description = "Document classification and OCR service for the Document Pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "transformers>=4.53.0",
    "torch>=2.6.0",
    "accelerate>=1.3.0",
    "sentencepiece>=0.2.0",
    "protobuf>=5.29.0",
    "Pillow>=11.1.0",
    "PyMuPDF>=1.25.0",
    "pydantic-settings>=2.7.0",
    "paddleocr>=2.7,<3",
    "paddlepaddle>=2.6,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not gpu'"
markers = [
    "gpu: requires CUDA GPU and downloaded model weights (deselected by default, run with: pytest -m gpu)",
]
